bring DRY level back up to 100%
