+++ package.json
@@ -15,7 +15,7 @@
     "lint": "eslint . --ext .ts,.tsx --fix",
     "folderslint": "folderslint",
     "prepare": "husky install",
-    "dry-test": "jscpd ./src --threshold=1",
+    "dry-test": "jscpd ./src --threshold=0",
     "generate-dry-report": "jscpd ./src --reporters=html",
     "mod": "node generateModTemplate.js",
     "ctm": "node changeToModTemplate.js && yarn format && yarn install && rm -r ./changeToModTemplate.js"
