+++ src/pages/App.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import { Button, VStack } from '@chakra-ui/react'
 import { useNavigate } from 'react-router-dom'
 import { useSelector, useDispatch } from 'react-redux'
@@ -6,45 +6,28 @@ import { DndContext, closestCenter, useSensors, KeyboardSensor, PointerSensor, u
 import { arrayMove, SortableContext, sortableKeyboardCoordinates, verticalListSortingStrategy } from '@dnd-kit/sortable'
 import { DragEndEvent } from '@dnd-kit/core/dist/types'
 
-import { selectMods, selectPWA, selectSelectedMods } from '~store'
+import { selectPWA } from '~store'
 import { setDeferredPrompt } from '~slices/pwaSlice'
 import type { AppDispatch } from '~store'
 import { Layout } from '~components/Layout'
-import { getMods } from '~slices/modSlice'
 import { ModBtnLink } from '~components/ModBtnLink'
-import { getSelectedMods, setSelectedMods } from '~slices/selectedModSlice'
+import { setSelectedMods } from '~slices/selectedModSlice'
 import { Header } from '~components/Header'
 import '../global.css'
+import { sortedModHook } from '../hooks/sortedModHook'
 
 export const App = (): JSX.Element => {
     const pwa = useSelector(selectPWA)
     const dispatch = useDispatch<AppDispatch>()
     const navigate = useNavigate()
-    const mods = useSelector(selectMods)
-    const selectedMods = useSelector(selectSelectedMods)
     const sensors = useSensors(
         useSensor(PointerSensor),
         useSensor(KeyboardSensor, {
             coordinateGetter: sortableKeyboardCoordinates,
         }),
     )
-    const [sortedMods, setSortedMods] = useState<Mod[]>([])
     const [isSortingMods, setIsSortingMods] = useState(false)
-
-    useEffect(() => {
-        dispatch(getSelectedMods())
-        dispatch(getMods())
-    }, [])
-
-    useEffect(() => {
-        const modsSorted: Mod[] = []
-        selectedMods.data.forEach((modId) => {
-            mods.data.find((mod) => {
-                if (mod.id === modId) modsSorted.push(mod)
-            })
-        })
-        setSortedMods(modsSorted)
-    }, [mods, selectedMods])
+    const { sortedMods, setSortedMods } = sortedModHook()
 
     const installHandler = async (e: React.MouseEvent<HTMLButtonElement>) => {
         if (pwa.deferredPrompt !== null) {
@@ -64,7 +47,7 @@ export const App = (): JSX.Element => {
                 const oldIndex = mods.findIndex((mod) => mod.id === active.id)
                 const newIndex = mods.findIndex((mod) => mod.id === over.id)
                 const sortedMods = arrayMove(mods, oldIndex, newIndex)
-                dispatch(setSelectedMods(sortedMods.map((mod) => mod.id)))
+                dispatch(setSelectedMods(sortedMods.map((mod: Mod) => mod.id)))
                 return sortedMods
             })
         }
